+++ tests/integration/test_various.py
@@ -14,7 +14,7 @@ def test_issue104() -> None:
     platform = Platform.CISCO_NXOS
     running_config = get_hconfig(get_hconfig_driver(platform), running_config_raw)
     generated_config = get_hconfig(get_hconfig_driver(platform), generated_config_raw)
-    remediation_config = running_config.config_to_get_to(generated_config)
+    remediation_config = running_config.remediation(generated_config)
     expected_rem_lines = {
         "no tacacs-server deadtime 3",
         "no tacacs-server host 192.168.1.99 key 7 Test12345",
@@ -22,6 +22,6 @@ def test_issue104() -> None:
         "tacacs-server host 192.168.100.98 key 0 test135 timeout 3",
     }
     remediation_lines = {
-        line.cisco_style_text() for line in remediation_config.all_children()
+        line.indented_text() for line in remediation_config.all_children()
     }
     assert expected_rem_lines == remediation_lines