+++ tests/unit/test_workflows.py
@@ -66,7 +66,7 @@ def test_rollback_config_reverts_changes(wfr: WorkflowRemediation) -> None:
     # Test if rollback config correctly represents changes needed to revert generated to running
     rollback_config = wfr.rollback_config
     rollback_text = "\n".join(
-        line.cisco_style_text() for line in rollback_config.all_children_sorted()
+        line.indented_text() for line in rollback_config.all_children_sorted()
     )
     expected_text = "no vlan 4\nno interface Vlan4\nvlan 3\n  name switch_mgmt_10.0.4.0/24\ninterface Vlan2\n  no mtu 9000\n  no ip access-group TEST in\n  shutdown\ninterface Vlan3\n  description switch_mgmt_10.0.4.0/24\n  ip address 10.0.4.1 255.255.0.0"
     assert rollback_text == expected_text